feat: Support attaching metadata to release ancestry links
BREAKING CHANGE: The `parent_versions` list in `seano-config.yaml`, and the
`before` and `after` lists on releases are now lists of dictionaries, not
lists of strings.  This primarily impacts seano views.  Old schema in
`seano-config.yaml` is automatically converted to the new schema on-the-fly.

As a brief example, within a seano view, if you currently do this:

    list_of_parent_releases = release['after']

You will need to change it to this:

    list_of_parent_releases = [x['name'] for x in release['after']]

For more information, refer to the new docs in this commit.

This commit exists for posterity only; data unrelated to the history of the QA Notes Seano formatter plugin has been removed.
Original repository: zarf.git
Original commit ID: 7331014456824a26704776076c403598094aec52